fix(audit): scrolling height verbs — gap accounting, override drops, tabbed ownership
Remediation for the code audit of #1060. Eight read-only audit passes over the
frozen branch found 40 findings; these are the code fixes.

Engine, all in the window-height verbs the PR added:

- expandActiveWindowToAvailableHeight subtracted the inter-tile gaps twice. The
  budget from activeColumnCrossBudgetPx has already removed them and relayout
  keeps gaps outside the tile rects, so seeding `taken` with gap*(tiles-1)
  stranded exactly one gap's worth on every press and then reported "nothing
  left" while it was free. The width twin's seed is correct there because it
  measures against the raw viewport. Seed dropped; the ops expectation moves
  520 -> 530 and now asserts the column tiles its budget exactly.

- The same verb measured a maximize-to-edges column without dropping the
  override first. Under the flag relayout resolves every tile as a share of the
  RAW cross extent, so the walk compared raw-area shares against a gapped
  budget and refused permanently. It now drops the flag before measuring and
  reports the drop from every bail, the way expandActiveColumnToAvailableWidth
  does.

- The same verb routed a tabbed column into the maximize toggle BEFORE its
  "already full" early-out, so a grow press on a full tab reached the toggle's
  un-maximize arm and discarded the intent while reporting success. The
  early-out is hoisted above the tabbed route and measures in column space.

- toggleMaximizeActiveWindowHeight and minimizeActiveWindowHeight gated the
  override drop on a value change. setActiveWindowHeight clears it
  unconditionally while set, and says why: under the override the stored
  intents are not rendered, so a result equal to the stored value is still a
  real change. The equality test made a lone Auto tile in a flagged column
  refuse forever, on the one press whose job is to leave the override.

- toggleMaximizeActiveWindowHeight only recognised a Fixed intent as maximized,
  so a Preset resolving to the budget was rewritten to a Fixed of the same
  extent: success reported, nothing moved, proportional anchor gone.

- minimizeActiveWindowHeight capped its target on the inside but not the
  outside, so a client floor above the column budget produced a target no
  relayout can render and every press reported success with nothing moving.

- minimizeActiveWindowHeight claimed tabbed extent ownership and then bailed on
  a rendered-pixel compare without writing the intent. claimTabbedHeightOwnership
  moves the owner and writes nothing else, and tabbedColumnCrossPx resolves an
  Auto owner to the whole work area — so a press asking to shrink grew the
  column to full height. The write now precedes the bail.

- reconcileWindowSize overwrote the Fixed(budget) intent that is the toggle's
  only un-maximize memory. A client with size increments acks a cross extent
  short of the one applied, which made maximize one-way for the life of the
  window. A short ack against a stored maximize intent is now treated as the
  client declining, not choosing. columnCrossBudgetPx is split out of
  activeColumnCrossBudgetPx because reconcile answers for the resized column.

- equalizeWindowHeights emitted the width verb's "equalize" OSD token, which
  the OSD renders as "Column widths equalized" for a verb that changes no
  column's width. It now emits equalize_heights, with a matching arm.

- scrollstrip_sizing.cpp had crossed the 1150 ceiling from under it (1147 ->
  1376). The height-verb family moves to scrollstrip_sizing_height.cpp; the two
  files are 933 and 573.

Tests. Two new slots pin the tabbed ownership bug and the tabbed grow refusal,
both confirmed to fail against the unfixed engine before the fix went in. The
three whole-size height verbs join the maximize-to-edges clearing enumeration,
whose header claims to drive every clearing site and did not. Six advertised
chords are pinned by value, and the one Settings slot without an
IsolatedConfigGuard gets one.

Comments and docs corrected where they were false rather than merely terse: the
chord banner claimed a rule three of seven height verbs follow, a
cross-reference named an accessor that does not exist, and the externally-owned
chord table is now explicit that the Meta+Ctrl+Alt space is unsurveyed rather
than clear.
