+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Versioning follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- **Scrolling window heights now have the same four shortcuts column widths always had**: a column could be maximized, grown into the empty space beside it, shrunk to its narrowest preset and equalized with its neighbours, but a window inside a column could only be grown, shrunk and cycled through the presets. Maximize Window Height, Grow Window into Empty Space and Minimize Window Height are new, and the old Reset Window Heights is now Equalize Window Heights, which is what it did and what its width twin is called. The four height shortcuts are the width chord with Ctrl added, so Meta+Alt+F maximizes the column and Meta+Ctrl+Alt+F maximizes the window in it, and the same holds for E and Shift+E. Equalize Window Heights keeps Meta+Ctrl+Alt+T. Since the old Reset Window Heights setting was renamed, a shortcut you had set on it yourself goes back to the default.
+- **Scrolling window heights now have the same four shortcuts column widths have**: a column could be maximized, grown into the empty space beside it, shrunk to its smallest preset and equalized with its neighbours, but a window inside a column could only be grown, shrunk and cycled through the presets. Maximize Window Height, Grow Window into Empty Space and Minimize Window Height are new, and the old Reset Window Heights is now Equalize Window Heights, which is what it did and what its width twin is called. The three new ones are their width chord with Ctrl added, so Meta+Alt+F maximizes the column and Meta+Ctrl+Alt+F maximizes the window in it, and E and Shift+E follow the same rule. Equalize Window Heights stays on Meta+Ctrl+Alt+T, beside the other verbs that re-flow the strip. Since the old Reset Window Heights setting was renamed, a shortcut you had set on it yourself goes back to the default. ([#1060](https://github.com/fuddlesworth/PlasmaZones/pull/1060))
 
 ### Changed
