+++ libs/phosphor-scroll-engine/CMakeLists.txt
@@ -34,6 +34,7 @@ add_library(PhosphorScrollEngine SHARED
     src/scrollstrip_navigation.cpp
     src/scrollstrip_relayout.cpp
     src/scrollstrip_sizing.cpp
+    src/scrollstrip_sizing_height.cpp
     src/scrollengine/engine_core.cpp
     src/scrollengine/engine_overrides.cpp
     src/scrollengine/engine_context.cpp
