+++ libs/phosphor-scroll-engine/src/scrollengine/engine_verbs.cpp
@@ -454,10 +454,15 @@ void ScrollEngine::expandWindowToAvailableHeight(const QString& screenId)
 
 void ScrollEngine::equalizeWindowHeights(const QString& screenId)
 {
-    // "equalize" in the reason slot, equalizeVisibleColumnWidths' reason: the
-    // OSD's generic resize copy names one window, and this verb rewrote a
-    // whole column's worth. The failure token stays the shared no_target.
-    P_SCROLL_VERB(screenId, state->strip().equalizeActiveColumnHeights(), "resize", false, QStringLiteral("equalize"));
+    // "equalize_heights" in the reason slot, the height sibling of
+    // equalizeVisibleColumnWidths' "equalize": the OSD's generic resize copy
+    // names one window, and this verb rewrote a whole column's worth. The two
+    // tokens must stay DISTINCT — the OSD renders each by name, so borrowing
+    // the width verb's token announced "Column widths equalized" for a verb
+    // that changed no column's width. The failure token stays the shared
+    // no_target.
+    P_SCROLL_VERB(screenId, state->strip().equalizeActiveColumnHeights(), "resize", false,
+                  QStringLiteral("equalize_heights"));
 }
 
 void ScrollEngine::centerVisibleColumns(const QString& screenId)
